db,experiments: move modules at the top level
