+++ src/lib.rs
@@ -81,12 +81,14 @@ pub mod log;
 pub mod util;
 pub mod config;
 pub mod crates;
+mod db;
 pub mod dirs;
 mod dl;
 pub mod docker;
 pub mod errors;
 pub mod ex;
 pub mod ex_run;
+mod experiments;
 mod file;
 mod gh;
 mod git;
